requirements.txt
absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
jax==0.6.2
jaxlib==0.6.2
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mediapipe==0.10.14
ml_dtypes==0.5.1
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
openai==1.97.0
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
ruff==0.12.3
safehttpx==0.1.6
scipy==1.16.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
starlette==0.47.1
sympy==1.14.0
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
